% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_taxa.R
\name{classify_taxa}
\alias{classify_taxa}
\title{Traditional categories of microbial taxa based on the relative abundance}
\usage{
classify_taxa(x, t_lower = 1e-04, t_upper = 0.01, method = "0")
}
\arguments{
\item{x}{The n by m matrix of microbial abundance.}

\item{t_lower}{The lower threshold to separate rare and intermediate taxa.}

\item{t_upper}{The upper threshold to separate intermediate and abundant taxa.}

\item{method}{The method index which can be 'a', 'b', 'c', or '0'.}
}
\value{
A list of microbial taxa categories for a given method.
}
\description{
This function classify microbial taxa into six categories -
1) RT: rare taxa;
2) IT: intermediate taxa;
3) AT: abundant taxa;
4) CRT: conditionally rare taxa;
5) CAT: conditionally abundant taxa;
6) CRAT: conditionally rare or abundant taxa -
by following four popular methods summerized in Tang et al. (2022) including -
1) Method 0 (Dai et al. 2016);
2) Method a (Campbell et al. 2011);
3) Method b (Logares et al. 2014);
4) Method c (Dai et al. 2016)
}
\examples{
# import data
data(hzmicrobe)
# try method 0 where the lower and upper thresholds can be adjusted
category_method_0 <- classify_taxa(mic_bay, t_l=0.0001, t_u=0.01, method='0')
# print abundant taxa id
print(category_method_0$AT)
# similar for printing out other category groups
# try method a where the lower and upper thresholds are fixed
category_method_a <- classify_taxa(mic_bay, method='a')
print(category_method_a$AT)
# try method b
category_method_b <- classify_taxa(mic_bay, method='b')
print(category_method_b$AT)
# try method c
category_method_c <- classify_taxa(mic_bay, method='c')
print(category_method_c$AT)


}
