% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dispersion.R
\name{is_dispersion}
\alias{is_dispersion}
\title{Dispersion test}
\usage{
is_dispersion(v)
}
\arguments{
\item{v}{The m-length vector, denotying the abundance of a specific microbe.}
}
\value{
True or False.
}
\description{
This function examines the dispersion pattern.
}
\examples{
# import data
data(hzmicrobe)
is_dispersion(as.numeric(mic_bay[1,]))
is_dispersion(as.numeric(mic_era[12,]))


}
