% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_overdispersion.R
\name{is_overdispersion}
\alias{is_overdispersion}
\title{Over-dispersion test}
\usage{
is_overdispersion(v, alpha = 0.05)
}
\arguments{
\item{v}{An m-length vector, denoting the abundance of a specific microbe.}
}
\value{
True or False.
}
\description{
This function diagnose the over-dispersion generally for microbial abundance data.
}
\examples{
# import data
data(hzmicrobe)
is_overdispersion(as.numeric(mic_bay[1,]))
is_overdispersion(as.numeric(mic_era[12,]))


}
