% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_zero_infla.R
\name{is_zero_infla}
\alias{is_zero_infla}
\title{Zero-inflation test}
\usage{
is_zero_infla(v, n_sample = NULL, replicates = NULL, max_prop_zero = 2/3)
}
\arguments{
\item{v}{An m-length vector, denoting the abundance of a specific microbe.}

\item{n_sample}{The numeric number of sample size.}

\item{replicates}{The numeric number of replicates per sample site.}

\item{max_prop_zero}{The maximum acceptable value of zero proportions per sample site.}
}
\value{
True or False.
}
\description{
This function detects the existence of inflated zeros.
}
\examples{
# import data
data(hzmicrobe)
is_zero_infla(as.numeric(mic_bay[1,]), n_sample=10, replicates=3)
is_zero_infla(as.numeric(mic_era[12,]), n_sample=12, replicates=3)


}
