% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_simple.R
\name{pca_mic}
\alias{pca_mic}
\title{A simpe algorithm for principle components analysis}
\usage{
pca_mic(x, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{The n by m matrix of microbial abundance with columns denoting sample sites.}
}
\value{
A list of principle components, loadings, and proportion of variance explained by each pc.
}
\description{
This function conducts PCA for microbial abundance data.
}
\examples{
# import data
data(hzmicrobe)
pca_bay <- pca_mic(mic_bay)


}
