% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_boundary.R
\name{refine_mu_k}
\alias{refine_mu_k}
\title{Refine the boundary between sub-communities.}
\usage{
refine_mu_k(x)
}
\arguments{
\item{x}{The $n$ by $m$ matrix of microbial abundance.}
}
\value{
A list of microbes' id for dispersal vanguards and laggards.
}
\description{
This function computes a model-driven testable boundary between dispersal vanguards and laggards.
It assigns microbes in the temporary transition region into either mu-wing or k-wing.
}
\examples{
# import data
data(hzmicrobe)
id_lis <- refine_mu_k(mic_bay)

}
